GradientPicker.Overlay: handle-only overlay for in-canvas editing
Extracts the visual handle layer of `<GradientPicker.Area>` into its
own composable part. Useful when a consumer wants to let users edit
the gradient directly on their own object on a canvas: drop the
overlay over any element, gradient handles align to that element's box.

- New `<GradientPicker.Overlay>` part owns all of the handle math +
  drag + keyboard logic + ResizeObserver + containerWidth publishing
  + radii-snapshot-on-drag-end. Root div is
  `absolute inset-0 pointer-events-none` so empty regions pass clicks
  through to whatever sits beneath. Handle buttons are
  `pointer-events-auto` to keep receiving input. Works identically
  for linear (endpoints + middle stops), radial (center + edge),
  and conic (center + dial).
- `<GradientPicker.Area>` is now a thin wrapper: relative container,
  transparency checkerboard, painted gradient, then `<Overlay />` on
  top. Same external props (`height`, `conicDialRadius`); no
  behavior change for consumers using Area.
- Registered `overlay.tsx` in `registry.json` and re-exported via the
  GradientPicker barrel.
- Playground gets a new `Overlay (demo)` toggle that renders a sample
  canvas object with the live gradient applied and the overlay sitting
  on top. The code generator emits the matching markup for copy-paste.
- Docs get a dedicated section (`#gradient-overlay`) with a runnable
  example showing the Overlay over a consumer-owned div, plus prose
  on the host container constraints (positioning context, no
  transforms, pointer-event semantics).
- TOC gains anchors for Area, Overlay, and radial size keywords.
