+++ registry.json
@@ -158,6 +158,11 @@
           "target": "components/ui/fill-picker/parts/gradient/bar.tsx"
         },
         {
+          "path": "registry/new-york/color-picker/parts/gradient/overlay.tsx",
+          "type": "registry:ui",
+          "target": "components/ui/fill-picker/parts/gradient/overlay.tsx"
+        },
+        {
           "path": "registry/new-york/color-picker/parts/gradient/area.tsx",
           "type": "registry:ui",
           "target": "components/ui/fill-picker/parts/gradient/area.tsx"
